// apps/api/src/generated/client/enums.ts
export const PatientStatus = {
  REGISTERED: 'REGISTERED',
  APPOINTMENT_SCHEDULED: 'APPOINTMENT_SCHEDULED',
  CHECKED_IN: 'CHECKED_IN',
  WAITING: 'WAITING',
  CALLED: 'CALLED',
  CONSULTATION: 'CONSULTATION',
  LABORATORY: 'LABORATORY',
  PHARMACY: 'PHARMACY',
  BILLING: 'BILLING',
  ADMITTED: 'ADMITTED',
  DISCHARGED: 'DISCHARGED'
} as const

export type PatientStatus = (typeof PatientStatus)[keyof typeof PatientStatus]
// ...
export const BedStatus = {
  AVAILABLE: 'AVAILABLE',
  OCCUPIED: 'OCCUPIED',
  RESERVED: 'RESERVED',
  CLEANING: 'CLEANING',
  MAINTENANCE: 'MAINTENANCE',
  BLOCKED: 'BLOCKED'
} as const

export type BedStatus = (typeof BedStatus)[keyof typeof BedStatus]


export const BedType = {
  STANDARD: 'STANDARD',
  ELECTRIC: 'ELECTRIC',
  ICU: 'ICU',
  PEDIATRIC: 'PEDIATRIC',
  BARIATRIC: 'BARIATRIC'
} as const

export type BedType = (typeof BedType)[keyof typeof BedType]


export const AdmissionStatus = {
  ADMITTED: 'ADMITTED',
  TRANSFERRED: 'TRANSFERRED',
  DISCHARGED: 'DISCHARGED'
} as const

export type AdmissionStatus = (typeof AdmissionStatus)[keyof typeof AdmissionStatus]


export const AdmissionType = {
  EMERGENCY: 'EMERGENCY',
  ELECTIVE: 'ELECTIVE',
  TRANSFER: 'TRANSFER',
  DAY_CARE: 'DAY_CARE'
} as const

export type AdmissionType = (typeof AdmissionType)[keyof typeof AdmissionType]


export const DischargeType = {
  NORMAL: 'NORMAL',
  AGAINST_MEDICAL_ADVICE: 'AGAINST_MEDICAL_ADVICE',
  TRANSFER: 'TRANSFER',
  DECEASED: 'DECEASED'
} as const

export type DischargeType = (typeof DischargeType)[keyof typeof DischargeType]